requirements.txt
openai>=1.0.0
requests>=2.28.0
pydantic>=1.10.0
pandas>=1.3.0
python-dotenv>=0.21.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
joblib>=1.3.0
